reflect preferredSize change to long
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@802766 13f79535-47bb-0310-9956-ffa450edef68

+++ kahadb/src/main/java/org/apache/kahadb/journal/ReadOnlyDataFile.java
@@ -26,7 +26,7 @@ import java.io.RandomAccessFile;
  */
 public class ReadOnlyDataFile extends DataFile {
 
-    ReadOnlyDataFile(File file, int number, int preferedSize) {
+    ReadOnlyDataFile(File file, int number, long preferedSize) {
         super(file, number, preferedSize);
     }
